@@ -3,31 +3,28 @@
   "version": "0.0.1",
   "private": true,
   "dependencies": {
-    "@capacitor/core": "1.3.0",
-    "@ionic/react": "^6.1.15",
-    "@ionic/react-router": "^6.1.15",
-    "@types/jest": "24.0.18",
-    "@types/node": "12.7.5",
-    "@types/react": "^16.9.43",
-    "@types/react-dom": "^16.9.8",
-    "@types/react-router": "^5.1.8",
-    "@types/react-router-dom": "^5.1.5",
+    "@capacitor/core": "^4.6.1",
+    "@capacitor/preferences": "^4.0.2",
+    "@ionic/react": "^7.0.0",
+    "@ionic/react-router": "^7.0.0",
     "date-fns": "^2.25.0",
     "ionicons": "^5.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^3.4.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router": "^5.3.4",
+    "react-router-dom": "^5.3.4",
     "reselect": "^4.0.0",
-    "sass": "^1.43.4",
-    "swiper": "^7.2.0"
+    "sass": "^1.59.3",
+    "swiper": "^9.1.1"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "start": "npm run dev",
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview",
+    "format": "prettier --write \"src/**/*.{js,jsx,ts,tsx,json,md}\"",
+    "precommit": "npm run format && lint-staged",
+    "prepare": "husky install"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -40,9 +37,25 @@
   ],
   "description": "An Ionic project",
   "devDependencies": {
-    "@capacitor/cli": "1.3.0",
+    "@capacitor/cli": "^4.6.1",
     "@testing-library/react": "^9.3.1",
     "@types/googlemaps": "^3.38.0",
-    "typescript": "~3.8.0"
+    "@types/jest": "24.0.18",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
+    "@types/react-router": "^5.1.20",
+    "@types/react-router-dom": "^5.3.3",
+    "@vitejs/plugin-react": "^3.1.0",
+    "husky": "^8.0.3",
+    "lint-staged": "^13.2.0",
+    "prettier": "^2.8.6",
+    "typescript": "^4.9.3",
+    "vite": "^4.2.0"
+  },
+  "lint-staged": {
+    "src/**/*.{js,jsx,ts,tsx,json,md}": [
+      "prettier --write",
+      "git add"
+    ]
   }
 }
