@@ -3,10 +3,9 @@ import { IonList, IonItem, IonLabel } from '@ionic/react';
 
 interface AboutPopoverProps {
   dismiss: () => void;
-};
+}
 
-const AboutPopover: React.FC<AboutPopoverProps> = ({dismiss}) => {
-
+const AboutPopover: React.FC<AboutPopoverProps> = ({ dismiss }) => {
   const close = (url: string) => {
     window.open(url, '_blank');
     dismiss();
@@ -14,23 +13,32 @@ const AboutPopover: React.FC<AboutPopoverProps> = ({dismiss}) => {
 
   return (
     <IonList>
-      <IonItem button onClick={() => close('https://ionicframework.com/getting-started')}>
+      <IonItem button onClick={() => close('https://ionicframework.com/docs')}>
         <IonLabel>Learn Ionic</IonLabel>
       </IonItem>
-      <IonItem button onClick={() => close('https://ionicframework.com/docs/react')}>
+      <IonItem
+        button
+        onClick={() => close('https://ionicframework.com/docs/react')}
+      >
         <IonLabel>Documentation</IonLabel>
       </IonItem>
-      <IonItem button onClick={() => close('https://showcase.ionicframework.com')}>
+      <IonItem
+        button
+        onClick={() => close('https://showcase.ionicframework.com')}
+      >
         <IonLabel>Showcase</IonLabel>
       </IonItem>
-      <IonItem button onClick={() => close('https://github.com/ionic-team/ionic')}>
+      <IonItem
+        button
+        onClick={() => close('https://github.com/ionic-team/ionic-framework')}
+      >
         <IonLabel>GitHub Repo</IonLabel>
       </IonItem>
       <IonItem button onClick={dismiss}>
         <IonLabel>Support</IonLabel>
       </IonItem>
-    </IonList >
-  )
-}
+    </IonList>
+  );
+};
 
 export default AboutPopover;
