@@ -7,12 +7,16 @@ interface StateProps {
 }
 
 const HomeOrTutorial: React.FC<StateProps> = ({ hasSeenTutorial }) => {
-  return hasSeenTutorial ? <Redirect to="/tabs/schedule" /> : <Redirect to="/tutorial" />
+  return hasSeenTutorial ? (
+    <Redirect to="/tabs/schedule" />
+  ) : (
+    <Redirect to="/tutorial" />
+  );
 };
 
 export default connect<{}, StateProps, {}>({
   mapStateToProps: (state) => ({
-    hasSeenTutorial: state.user.hasSeenTutorial
+    hasSeenTutorial: state.user.hasSeenTutorial,
   }),
-  component: HomeOrTutorial
+  component: HomeOrTutorial,
 });
