@@ -2,8 +2,8 @@ import React, { useRef, useEffect } from 'react';
 import { Location } from '../models/Location';
 
 interface MapProps {
-  locations: Location[]
-  mapCenter: Location
+  locations: Location[];
+  mapCenter: Location;
 }
 
 const Map: React.FC<MapProps> = ({ mapCenter, locations }) => {
@@ -11,13 +11,12 @@ const Map: React.FC<MapProps> = ({ mapCenter, locations }) => {
   const map = useRef<google.maps.Map>();
 
   useEffect(() => {
-
     map.current = new google.maps.Map(mapEle.current, {
       center: {
         lat: mapCenter.lat,
-        lng: mapCenter.lng
+        lng: mapCenter.lng,
       },
-      zoom: 16
+      zoom: 16,
     });
 
     addMarkers();
@@ -31,26 +30,23 @@ const Map: React.FC<MapProps> = ({ mapCenter, locations }) => {
     function addMarkers() {
       locations.forEach((markerData) => {
         let infoWindow = new google.maps.InfoWindow({
-          content: `<h5>${markerData.name}</h5>`
+          content: `<h5>${markerData.name}</h5>`,
         });
-  
+
         let marker = new google.maps.Marker({
           position: new google.maps.LatLng(markerData.lat, markerData.lng),
           map: map.current!,
-          title: markerData.name
+          title: markerData.name,
         });
-  
+
         marker.addListener('click', () => {
           infoWindow.open(map.current!, marker);
         });
       });
     }
-
   }, [mapCenter, locations]);
 
-  return (
-    <div ref={mapEle} className="map-canvas"></div>
-  );
-}
+  return <div ref={mapEle} className="map-canvas"></div>;
+};
 
 export default Map;
