@@ -1,5 +1,13 @@
-import { IonItemDivider, IonItemGroup, IonLabel, IonList, IonListHeader, IonAlert, AlertButton } from '@ionic/react';
-import React, { useState, useCallback } from 'react';
+import {
+  IonItemDivider,
+  IonItemGroup,
+  IonLabel,
+  IonList,
+  IonListHeader,
+  IonAlert,
+  AlertButton,
+} from '@ionic/react';
+import React, { useState, useCallback, useRef, useEffect } from 'react';
 import { Schedule, Session } from '../models/Schedule';
 import SessionListItem from './SessionListItem';
 import { connect } from '../data/connect';
@@ -20,39 +28,55 @@ interface DispatchProps {
   removeFavorite: typeof removeFavorite;
 }
 
-interface SessionListProps extends OwnProps, StateProps, DispatchProps { };
+interface SessionListProps extends OwnProps, StateProps, DispatchProps {}
 
-const SessionList: React.FC<SessionListProps> = ({ addFavorite, removeFavorite, favoriteSessions, hide, schedule, listType }) => {
-
+const SessionList: React.FC<SessionListProps> = ({
+  addFavorite,
+  removeFavorite,
+  favoriteSessions,
+  hide,
+  schedule,
+  listType,
+}) => {
+  const scheduleListRef = useRef<HTMLIonListElement>(null);
   const [showAlert, setShowAlert] = useState(false);
   const [alertHeader, setAlertHeader] = useState('');
-  const [alertButtons, setAlertButtons] = useState<(AlertButton | string)[]>([]);
+  const [alertMessage, setAlertMessage] = useState('');
+  const [alertButtons, setAlertButtons] = useState<(AlertButton | string)[]>(
+    []
+  );
 
-  const handleShowAlert = useCallback((header: string, buttons: AlertButton[]) => {
-    setAlertHeader(header);
-    setAlertButtons(buttons);
-    setShowAlert(true);
-  }, []);
+  const handleShowAlert = useCallback(
+    (header: string, message: string, buttons: AlertButton[]) => {
+      setAlertHeader(header);
+      setAlertMessage(message);
+      setAlertButtons(buttons);
+      setShowAlert(true);
+    },
+    []
+  );
+
+  useEffect(() => {
+    if (scheduleListRef.current) {
+      scheduleListRef.current.closeSlidingItems();
+    }
+  }, [hide]);
 
   if (schedule.groups.length === 0 && !hide) {
     return (
       <IonList>
-        <IonListHeader>
-          No Sessions Found
-        </IonListHeader>
+        <IonListHeader>No Sessions Found</IonListHeader>
       </IonList>
     );
   }
 
   return (
     <>
-      <IonList style={hide ? { display: 'none' } : {}}>
+      <IonList ref={scheduleListRef} style={hide ? { display: 'none' } : {}}>
         {schedule.groups.map((group, index: number) => (
           <IonItemGroup key={`group-${index}`}>
             <IonItemDivider sticky>
-              <IonLabel>
-                {group.time}
-              </IonLabel>
+              <IonLabel>{group.time}</IonLabel>
             </IonItemDivider>
             {group.sessions.map((session: Session, sessionIndex: number) => (
               <SessionListItem
@@ -71,6 +95,7 @@ const SessionList: React.FC<SessionListProps> = ({ addFavorite, removeFavorite, 
       <IonAlert
         isOpen={showAlert}
         header={alertHeader}
+        message={alertMessage}
         buttons={alertButtons}
         onDidDismiss={() => setShowAlert(false)}
       ></IonAlert>
@@ -80,11 +105,11 @@ const SessionList: React.FC<SessionListProps> = ({ addFavorite, removeFavorite, 
 
 export default connect<OwnProps, StateProps, DispatchProps>({
   mapStateToProps: (state) => ({
-    favoriteSessions: state.data.favorites
+    favoriteSessions: state.data.favorites,
   }),
-  mapDispatchToProps: ({
+  mapDispatchToProps: {
     addFavorite,
-    removeFavorite
-  }),
-  component: SessionList
+    removeFavorite,
+  },
+  component: SessionList,
 });
