@@ -1,10 +1,35 @@
 import React from 'react';
 
 import { getMode } from '@ionic/core';
-import { IonHeader, IonToolbar, IonButtons, IonButton, IonTitle, IonContent, IonList, IonListHeader, IonItem, IonLabel, IonCheckbox, IonFooter, IonIcon } from '@ionic/react';
-import { logoAngular, call, document, logoIonic, hammer, restaurant, cog, colorPalette, construct, compass } from 'ionicons/icons';
+import {
+  IonHeader,
+  IonToolbar,
+  IonButtons,
+  IonButton,
+  IonTitle,
+  IonContent,
+  IonList,
+  IonListHeader,
+  IonItem,
+  IonLabel,
+  IonCheckbox,
+  IonFooter,
+  IonIcon,
+} from '@ionic/react';
+import {
+  logoAngular,
+  call,
+  document,
+  logoIonic,
+  hammer,
+  restaurant,
+  cog,
+  colorPalette,
+  construct,
+  compass,
+} from 'ionicons/icons';
 
-import './SessionListFilter.css'
+import './SessionListFilter.css';
 
 import { connect } from '../data/connect';
 import { updateFilteredTracks } from '../data/sessions/sessions.actions';
@@ -14,8 +39,8 @@ interface OwnProps {
 }
 
 interface StateProps {
-  allTracks: string[],
-  filteredTracks: string[]
+  allTracks: string[];
+  filteredTracks: string[];
 }
 
 interface DispatchProps {
@@ -24,12 +49,17 @@ interface DispatchProps {
 
 type SessionListFilterProps = OwnProps & StateProps & DispatchProps;
 
-const SessionListFilter: React.FC<SessionListFilterProps> = ({ allTracks, filteredTracks, onDismissModal, updateFilteredTracks }) => {
+const SessionListFilter: React.FC<SessionListFilterProps> = ({
+  allTracks,
+  filteredTracks,
+  onDismissModal,
+  updateFilteredTracks,
+}) => {
   const ios = getMode() === 'ios';
 
   const toggleTrackFilter = (track: string) => {
     if (filteredTracks.indexOf(track) > -1) {
-      updateFilteredTracks(filteredTracks.filter(x => x !== track));
+      updateFilteredTracks(filteredTracks.filter((x) => x !== track));
     } else {
       updateFilteredTracks([...filteredTracks, track]);
     }
@@ -44,63 +74,65 @@ const SessionListFilter: React.FC<SessionListFilterProps> = ({ allTracks, filter
   };
 
   const iconMap: { [key: string]: any } = {
-    'Angular': logoAngular,
-    'Documentation': document,
-    'Food': restaurant,
-    'Ionic': logoIonic,
-    'Tooling': hammer,
-    'Design': colorPalette,
-    'Services': cog,
-    'Workshop': construct,
-    'Navigation': compass,
-    'Communication': call
-  }
+    Angular: logoAngular,
+    Documentation: document,
+    Food: restaurant,
+    Ionic: logoIonic,
+    Tooling: hammer,
+    Design: colorPalette,
+    Services: cog,
+    Workshop: construct,
+    Navigation: compass,
+    Communication: call,
+  };
 
   return (
     <>
       <IonHeader translucent={true} className="session-list-filter">
         <IonToolbar>
           <IonButtons slot="start">
-            { ios &&
-              <IonButton onClick={onDismissModal}>Cancel</IonButton>
-            }
-            { !ios &&
-              <IonButton onClick={handleDeselectAll}>Reset</IonButton>
-            }
+            {ios && <IonButton onClick={onDismissModal}>Cancel</IonButton>}
+            {!ios && <IonButton onClick={handleDeselectAll}>Reset</IonButton>}
           </IonButtons>
 
-          <IonTitle>
-            Filter Sessions
-          </IonTitle>
+          <IonTitle>Filter Sessions</IonTitle>
 
           <IonButtons slot="end">
-            <IonButton onClick={onDismissModal} strong>Done</IonButton>
+            <IonButton onClick={onDismissModal} strong>
+              Done
+            </IonButton>
           </IonButtons>
         </IonToolbar>
       </IonHeader>
 
       <IonContent className="session-list-filter">
-        <IonList lines={ ios ? 'inset' : 'full'}>
+        <IonList lines={ios ? 'inset' : 'full'}>
           <IonListHeader>Tracks</IonListHeader>
 
-          {allTracks.map((track, index) => (
+          {allTracks.map((track) => (
             <IonItem key={track}>
-              { ios &&
-                <IonIcon slot="start" icon={iconMap[track]} color="medium" />
-              }
-              <IonLabel>{track}</IonLabel>
+              {ios && (
+                <IonIcon
+                  slot="start"
+                  icon={iconMap[track]}
+                  color="medium"
+                  aria-hidden="true"
+                />
+              )}
               <IonCheckbox
-                onClick={() => toggleTrackFilter(track)}
+                onIonChange={() => toggleTrackFilter(track)}
                 checked={filteredTracks.indexOf(track) !== -1}
                 color="primary"
                 value={track}
-              ></IonCheckbox>
+              >
+                {track}
+              </IonCheckbox>
             </IonItem>
           ))}
         </IonList>
       </IonContent>
 
-      { ios &&
+      {ios && (
         <IonFooter>
           <IonToolbar>
             <IonButtons slot="start">
@@ -111,7 +143,7 @@ const SessionListFilter: React.FC<SessionListFilterProps> = ({ allTracks, filter
             </IonButtons>
           </IonToolbar>
         </IonFooter>
-      }
+      )}
     </>
   );
 };
@@ -119,10 +151,10 @@ const SessionListFilter: React.FC<SessionListFilterProps> = ({ allTracks, filter
 export default connect<OwnProps, StateProps, DispatchProps>({
   mapStateToProps: (state) => ({
     allTracks: state.data.allTracks,
-    filteredTracks: state.data.filteredTracks
+    filteredTracks: state.data.filteredTracks,
   }),
   mapDispatchToProps: {
-    updateFilteredTracks
+    updateFilteredTracks,
   },
-  component: SessionListFilter
-})
+  component: SessionListFilter,
+});
