@@ -1,9 +1,21 @@
-import { IonLoading, IonFab, IonFabButton, IonIcon, IonFabList } from "@ionic/react"
-import { shareSocial, logoVimeo, logoInstagram, logoTwitter, logoFacebook } from "ionicons/icons"
-import React, { useState } from "react"
+import {
+  IonLoading,
+  IonFab,
+  IonFabButton,
+  IonIcon,
+  IonFabList,
+} from '@ionic/react';
+import {
+  shareSocial,
+  logoVimeo,
+  logoInstagram,
+  logoTwitter,
+  logoFacebook,
+} from 'ionicons/icons';
+import React, { useState } from 'react';
 
 const ShareSocialFab: React.FC = () => {
-  const [loadingMessage, setLoadingMessage] = useState('')
+  const [loadingMessage, setLoadingMessage] = useState('');
   const [showLoading, setShowLoading] = useState(false);
 
   const openSocial = (network: string) => {
@@ -11,7 +23,7 @@ const ShareSocialFab: React.FC = () => {
     setShowLoading(true);
   };
 
-  return(
+  return (
     <>
       <IonLoading
         isOpen={showLoading}
@@ -28,7 +40,10 @@ const ShareSocialFab: React.FC = () => {
           <IonFabButton color="vimeo" onClick={() => openSocial('Vimeo')}>
             <IonIcon icon={logoVimeo} />
           </IonFabButton>
-          <IonFabButton color="instagram" onClick={() => openSocial('Instagram')}>
+          <IonFabButton
+            color="instagram"
+            onClick={() => openSocial('Instagram')}
+          >
             <IonIcon icon={logoInstagram} />
           </IonFabButton>
           <IonFabButton color="twitter" onClick={() => openSocial('Twitter')}>
@@ -40,7 +55,7 @@ const ShareSocialFab: React.FC = () => {
         </IonFabList>
       </IonFab>
     </>
-  )
+  );
 };
 
 export default ShareSocialFab;
