@@ -1,8 +1,15 @@
 import React from 'react';
 import { Session } from '../models/Schedule';
 import { Speaker } from '../models/Speaker';
-import { IonCard, IonCardHeader, IonItem, IonLabel, IonAvatar, IonCardContent, IonList } from '@ionic/react';
-
+import {
+  IonCard,
+  IonCardHeader,
+  IonItem,
+  IonLabel,
+  IonAvatar,
+  IonCardContent,
+  IonList,
+} from '@ionic/react';
 
 interface SpeakerItemProps {
   speaker: Speaker;
@@ -14,9 +21,15 @@ const SpeakerItem: React.FC<SpeakerItemProps> = ({ speaker, sessions }) => {
     <>
       <IonCard className="speaker-card">
         <IonCardHeader>
-          <IonItem button detail={false} lines="none" className="speaker-item" routerLink={`/tabs/speakers/${speaker.id}`}>
+          <IonItem
+            button
+            detail={false}
+            lines="none"
+            className="speaker-item"
+            routerLink={`/tabs/speakers/${speaker.id}`}
+          >
             <IonAvatar slot="start">
-              <img src={process.env.PUBLIC_URL + speaker.profilePic} alt="Speaker profile pic" />
+              <img src={speaker.profilePic} alt="Speaker profile pic" />
             </IonAvatar>
             <IonLabel>
               <h2>{speaker.name}</h2>
@@ -27,8 +40,12 @@ const SpeakerItem: React.FC<SpeakerItemProps> = ({ speaker, sessions }) => {
 
         <IonCardContent>
           <IonList lines="none">
-            {sessions.map(session => (
-              <IonItem detail={false} routerLink={`/tabs/speakers/sessions/${session.id}`} key={session.name}>
+            {sessions.map((session) => (
+              <IonItem
+                detail={false}
+                routerLink={`/tabs/speakers/sessions/${session.id}`}
+                key={session.name}
+              >
                 <IonLabel>
                   <h3>{session.name}</h3>
                 </IonLabel>
