@@ -1,5 +1,5 @@
-import React, { createContext, useReducer } from 'react';
-import { initialState, AppState, reducers } from './state'
+import React, { createContext, PropsWithChildren, useReducer } from 'react';
+import { initialState, AppState, reducers } from './state';
 
 export interface AppContextState {
   state: AppState;
@@ -8,19 +8,22 @@ export interface AppContextState {
 
 export const AppContext = createContext<AppContextState>({
   state: initialState,
-  dispatch: () => undefined
+  dispatch: () => undefined,
 });
 
-export const AppContextProvider: React.FC = (props => {
-
+export const AppContextProvider: React.FC<PropsWithChildren> = ({
+  children,
+}) => {
   const [store, dispatch] = useReducer(reducers, initialState);
 
   return (
-    <AppContext.Provider value={{
-      state: store,
-      dispatch
-    }}>
-      {props.children}
+    <AppContext.Provider
+      value={{
+        state: store,
+        dispatch,
+      }}
+    >
+      {children}
     </AppContext.Provider>
-  )
-});
+  );
+};
