@@ -1,7 +1,10 @@
 import { SessionsActions } from './sessions.actions';
 import { ConfState } from './conf.state';
 
-export const sessionsReducer = (state: ConfState, action: SessionsActions): ConfState => {
+export const sessionsReducer = (
+  state: ConfState,
+  action: SessionsActions
+): ConfState => {
   switch (action.type) {
     case 'set-conf-loading': {
       return { ...state, loading: action.isLoading };
@@ -10,10 +13,13 @@ export const sessionsReducer = (state: ConfState, action: SessionsActions): Conf
       return { ...state, ...action.data };
     }
     case 'add-favorite': {
-      return { ...state, favorites: [...(state.favorites), action.sessionId] };
+      return { ...state, favorites: [...state.favorites, action.sessionId] };
     }
     case 'remove-favorite': {
-      return { ...state, favorites: [...(state.favorites).filter(x => x !== action.sessionId)] };
+      return {
+        ...state,
+        favorites: [...state.favorites.filter((x) => x !== action.sessionId)],
+      };
     }
     case 'update-filtered-tracks': {
       return { ...state, filteredTracks: action.filteredTracks };
@@ -25,4 +31,4 @@ export const sessionsReducer = (state: ConfState, action: SessionsActions): Conf
       return { ...state, menuEnabled: action.menuEnabled };
     }
   }
-}
+};
