@@ -4,4 +4,4 @@ export interface UserState {
   darkMode: boolean;
   hasSeenTutorial: boolean;
   loading: boolean;
-};
+}
