@@ -1,11 +1,22 @@
 import React, { useState } from 'react';
-import { IonHeader, IonToolbar, IonTitle, IonContent, IonPage, IonButtons, IonMenuButton, IonList, IonItem, IonAlert } from '@ionic/react';
+import {
+  IonHeader,
+  IonToolbar,
+  IonTitle,
+  IonContent,
+  IonPage,
+  IonButtons,
+  IonMenuButton,
+  IonList,
+  IonItem,
+  IonAlert,
+} from '@ionic/react';
 import './Account.scss';
 import { setUsername } from '../data/user/user.actions';
 import { connect } from '../data/connect';
 import { RouteComponentProps } from 'react-router';
 
-interface OwnProps extends RouteComponentProps { }
+interface OwnProps extends RouteComponentProps {}
 
 interface StateProps {
   username?: string;
@@ -15,15 +26,14 @@ interface DispatchProps {
   setUsername: typeof setUsername;
 }
 
-interface AccountProps extends OwnProps, StateProps, DispatchProps { }
+interface AccountProps extends OwnProps, StateProps, DispatchProps {}
 
 const Account: React.FC<AccountProps> = ({ setUsername, username }) => {
-
   const [showAlert, setShowAlert] = useState(false);
 
   const clicked = (text: string) => {
     console.log(`Clicked ${text}`);
-  }
+  };
 
   return (
     <IonPage id="account-page">
@@ -36,19 +46,32 @@ const Account: React.FC<AccountProps> = ({ setUsername, username }) => {
         </IonToolbar>
       </IonHeader>
       <IonContent>
-        {username &&
-          (<div className="ion-padding-top ion-text-center">
-            <img src="https://www.gravatar.com/avatar?d=mm&s=140" alt="avatar" />
-            <h2>{ username }</h2>
+        {username && (
+          <div className="ion-padding-top ion-text-center">
+            <img
+              src="https://www.gravatar.com/avatar?d=mm&s=140"
+              alt="avatar"
+            />
+            <h2>{username}</h2>
             <IonList inset>
-              <IonItem onClick={() => clicked('Update Picture')}>Update Picture</IonItem>
-              <IonItem onClick={() => setShowAlert(true)}>Change Username</IonItem>
-              <IonItem onClick={() => clicked('Change Password')}>Change Password</IonItem>
-              <IonItem routerLink="/support" routerDirection="none">Support</IonItem>
-              <IonItem routerLink="/logout" routerDirection="none">Logout</IonItem>
+              <IonItem onClick={() => clicked('Update Picture')}>
+                Update Picture
+              </IonItem>
+              <IonItem onClick={() => setShowAlert(true)}>
+                Change Username
+              </IonItem>
+              <IonItem onClick={() => clicked('Change Password')}>
+                Change Password
+              </IonItem>
+              <IonItem routerLink="/support" routerDirection="none">
+                Support
+              </IonItem>
+              <IonItem routerLink="/logout" routerDirection="none">
+                Logout
+              </IonItem>
             </IonList>
-          </div>)
-        }
+          </div>
+        )}
       </IonContent>
       <IonAlert
         isOpen={showAlert}
@@ -59,16 +82,16 @@ const Account: React.FC<AccountProps> = ({ setUsername, username }) => {
             text: 'Ok',
             handler: (data) => {
               setUsername(data.username);
-            }
-          }
+            },
+          },
         ]}
         inputs={[
           {
             type: 'text',
             name: 'username',
             value: username,
-            placeholder: 'username'
-          }
+            placeholder: 'username',
+          },
         ]}
         onDidDismiss={() => setShowAlert(false)}
       />
@@ -78,10 +101,10 @@ const Account: React.FC<AccountProps> = ({ setUsername, username }) => {
 
 export default connect<OwnProps, StateProps, DispatchProps>({
   mapStateToProps: (state) => ({
-    username: state.user.username
+    username: state.user.username,
   }),
   mapDispatchToProps: {
     setUsername,
   },
-  component: Account
-})
+  component: Account,
+});
