@@ -1,5 +1,12 @@
-import React  from 'react';
-import { IonTabs, IonRouterOutlet, IonTabBar, IonTabButton, IonIcon, IonLabel } from '@ionic/react';
+import React from 'react';
+import {
+  IonTabs,
+  IonRouterOutlet,
+  IonTabBar,
+  IonTabButton,
+  IonIcon,
+  IonLabel,
+} from '@ionic/react';
 import { Route, Redirect } from 'react-router';
 import { calendar, location, informationCircle, people } from 'ionicons/icons';
 import SchedulePage from './SchedulePage';
@@ -9,10 +16,9 @@ import SessionDetail from './SessionDetail';
 import MapView from './MapView';
 import About from './About';
 
-interface MainTabsProps { }
+interface MainTabsProps {}
 
 const MainTabs: React.FC<MainTabsProps> = () => {
-
   return (
     <IonTabs>
       <IonRouterOutlet>
@@ -21,9 +27,21 @@ const MainTabs: React.FC<MainTabsProps> = () => {
           Using the render method prop cuts down the number of renders your components will have due to route changes.
           Use the component prop when your component depends on the RouterComponentProps passed in automatically.
         */}
-        <Route path="/tabs/schedule" render={() => <SchedulePage />} exact={true} />
-        <Route path="/tabs/speakers" render={() => <SpeakerList />} exact={true} />
-        <Route path="/tabs/speakers/:id" component={SpeakerDetail} exact={true} />
+        <Route
+          path="/tabs/schedule"
+          render={() => <SchedulePage />}
+          exact={true}
+        />
+        <Route
+          path="/tabs/speakers"
+          render={() => <SpeakerList />}
+          exact={true}
+        />
+        <Route
+          path="/tabs/speakers/:id"
+          component={SpeakerDetail}
+          exact={true}
+        />
         <Route path="/tabs/schedule/:id" component={SessionDetail} />
         <Route path="/tabs/speakers/sessions/:id" component={SessionDetail} />
         <Route path="/tabs/map" render={() => <MapView />} exact={true} />
