@@ -1,11 +1,29 @@
 import React, { useState, useRef } from 'react';
 
-import { IonToolbar, IonContent, IonPage, IonButtons, IonTitle, IonMenuButton, IonSegment, IonSegmentButton, IonButton, IonIcon, IonSearchbar, IonRefresher, IonRefresherContent, IonToast, IonModal, IonHeader, getConfig } from '@ionic/react';
+import {
+  IonToolbar,
+  IonContent,
+  IonPage,
+  IonButtons,
+  IonTitle,
+  IonMenuButton,
+  IonSegment,
+  IonSegmentButton,
+  IonButton,
+  IonIcon,
+  IonSearchbar,
+  IonRefresher,
+  IonRefresherContent,
+  IonToast,
+  IonModal,
+  IonHeader,
+  getConfig,
+} from '@ionic/react';
 import { options, search } from 'ionicons/icons';
 
 import SessionList from '../components/SessionList';
 import SessionListFilter from '../components/SessionListFilter';
-import './SchedulePage.scss'
+import './SchedulePage.scss';
 
 import ShareSocialFab from '../components/ShareSocialFab';
 
@@ -14,12 +32,12 @@ import { connect } from '../data/connect';
 import { setSearchText } from '../data/sessions/sessions.actions';
 import { Schedule } from '../models/Schedule';
 
-interface OwnProps { }
+interface OwnProps {}
 
 interface StateProps {
   schedule: Schedule;
   favoritesSchedule: Schedule;
-  mode: 'ios' | 'md'
+  mode: 'ios' | 'md';
 }
 
 interface DispatchProps {
@@ -28,7 +46,12 @@ interface DispatchProps {
 
 type SchedulePageProps = OwnProps & StateProps & DispatchProps;
 
-const SchedulePage: React.FC<SchedulePageProps> = ({ favoritesSchedule, schedule, setSearchText, mode }) => {
+const SchedulePage: React.FC<SchedulePageProps> = ({
+  favoritesSchedule,
+  schedule,
+  setSearchText,
+  mode,
+}) => {
   const [segment, setSegment] = useState<'all' | 'favorites'>('all');
   const [showSearchbar, setShowSearchbar] = useState<boolean>(false);
   const [showFilterModal, setShowFilterModal] = useState(false);
@@ -43,61 +66,66 @@ const SchedulePage: React.FC<SchedulePageProps> = ({ favoritesSchedule, schedule
     setTimeout(() => {
       ionRefresherRef.current!.complete();
       setShowCompleteToast(true);
-    }, 2500)
+    }, 2500);
   };
 
   return (
     <IonPage ref={pageRef} id="schedule-page">
       <IonHeader translucent={true}>
         <IonToolbar>
-          {!showSearchbar &&
+          {!showSearchbar && (
             <IonButtons slot="start">
               <IonMenuButton />
             </IonButtons>
-          }
-          {ios &&
-            <IonSegment value={segment} onIonChange={(e) => setSegment(e.detail.value as any)}>
-              <IonSegmentButton value="all">
-                All
-              </IonSegmentButton>
-              <IonSegmentButton value="favorites">
-                Favorites
-              </IonSegmentButton>
+          )}
+          {ios && (
+            <IonSegment
+              value={segment}
+              onIonChange={(e) => setSegment(e.detail.value as any)}
+            >
+              <IonSegmentButton value="all">All</IonSegmentButton>
+              <IonSegmentButton value="favorites">Favorites</IonSegmentButton>
             </IonSegment>
-          }
-          {!ios && !showSearchbar &&
-            <IonTitle>Schedule</IonTitle>
-          }
-          {showSearchbar &&
-            <IonSearchbar showCancelButton="always" placeholder="Search" onIonChange={(e: CustomEvent) => setSearchText(e.detail.value)} onIonCancel={() => setShowSearchbar(false)}></IonSearchbar>
-          }
+          )}
+          {!ios && !showSearchbar && <IonTitle>Schedule</IonTitle>}
+          {showSearchbar && (
+            <IonSearchbar
+              showCancelButton="always"
+              placeholder="Search"
+              onIonInput={(e: CustomEvent) => setSearchText(e.detail.value)}
+              onIonCancel={() => setShowSearchbar(false)}
+            ></IonSearchbar>
+          )}
 
           <IonButtons slot="end">
-            {!ios && !showSearchbar &&
+            {!ios && !showSearchbar && (
               <IonButton onClick={() => setShowSearchbar(true)}>
                 <IonIcon slot="icon-only" icon={search}></IonIcon>
               </IonButton>
-            }
-            {!showSearchbar &&
+            )}
+            {!showSearchbar && (
               <IonButton onClick={() => setShowFilterModal(true)}>
-                {mode === 'ios' ? 'Filter' : <IonIcon icon={options} slot="icon-only" />}
+                {mode === 'ios' ? (
+                  'Filter'
+                ) : (
+                  <IonIcon icon={options} slot="icon-only" />
+                )}
               </IonButton>
-            }
+            )}
           </IonButtons>
         </IonToolbar>
 
-        {!ios &&
+        {!ios && (
           <IonToolbar>
-            <IonSegment value={segment} onIonChange={(e) => setSegment(e.detail.value as any)}>
-              <IonSegmentButton value="all">
-                All
-              </IonSegmentButton>
-              <IonSegmentButton value="favorites">
-                Favorites
-              </IonSegmentButton>
+            <IonSegment
+              value={segment}
+              onIonChange={(e) => setSegment(e.detail.value as any)}
+            >
+              <IonSegmentButton value="all">All</IonSegmentButton>
+              <IonSegmentButton value="favorites">Favorites</IonSegmentButton>
             </IonSegment>
           </IonToolbar>
-        }
+        )}
       </IonHeader>
 
       <IonContent fullscreen={true}>
@@ -106,11 +134,18 @@ const SchedulePage: React.FC<SchedulePageProps> = ({ favoritesSchedule, schedule
             <IonTitle size="large">Schedule</IonTitle>
           </IonToolbar>
           <IonToolbar>
-            <IonSearchbar placeholder="Search" onIonChange={(e: CustomEvent) => setSearchText(e.detail.value)}></IonSearchbar>
+            <IonSearchbar
+              placeholder="Search"
+              onIonInput={(e: CustomEvent) => setSearchText(e.detail.value)}
+            ></IonSearchbar>
           </IonToolbar>
         </IonHeader>
 
-        <IonRefresher slot="fixed" ref={ionRefresherRef} onIonRefresh={doRefresh}>
+        <IonRefresher
+          slot="fixed"
+          ref={ionRefresherRef}
+          onIonRefresh={doRefresh}
+        >
           <IonRefresherContent />
         </IonRefresher>
 
@@ -127,7 +162,6 @@ const SchedulePage: React.FC<SchedulePageProps> = ({ favoritesSchedule, schedule
           hide={segment === 'favorites'}
         />
         <SessionList
-          // schedule={schedule}
           schedule={favoritesSchedule}
           listType={segment}
           hide={segment === 'all'}
@@ -137,16 +171,12 @@ const SchedulePage: React.FC<SchedulePageProps> = ({ favoritesSchedule, schedule
       <IonModal
         isOpen={showFilterModal}
         onDidDismiss={() => setShowFilterModal(false)}
-        swipeToClose={true}
         presentingElement={pageRef.current!}
       >
-        <SessionListFilter
-          onDismissModal={() => setShowFilterModal(false)}
-        />
+        <SessionListFilter onDismissModal={() => setShowFilterModal(false)} />
       </IonModal>
 
       <ShareSocialFab />
-
     </IonPage>
   );
 };
@@ -155,10 +185,10 @@ export default connect<OwnProps, StateProps, DispatchProps>({
   mapStateToProps: (state) => ({
     schedule: selectors.getSearchedSchedule(state),
     favoritesSchedule: selectors.getGroupedFavorites(state),
-    mode: getConfig()!.get('mode')
+    mode: getConfig()!.get('mode'),
   }),
   mapDispatchToProps: {
-    setSearchText
+    setSearchText,
   },
-  component: React.memo(SchedulePage)
+  component: React.memo(SchedulePage),
 });
