@@ -1,5 +1,18 @@
 import React from 'react';
-import { IonHeader, IonToolbar, IonContent, IonPage, IonButtons, IonBackButton, IonButton, IonIcon, IonText, IonList, IonItem, IonLabel } from '@ionic/react';
+import {
+  IonHeader,
+  IonToolbar,
+  IonContent,
+  IonPage,
+  IonButtons,
+  IonBackButton,
+  IonButton,
+  IonIcon,
+  IonText,
+  IonList,
+  IonItem,
+  IonLabel,
+} from '@ionic/react';
 import { connect } from '../data/connect';
 import { withRouter, RouteComponentProps } from 'react-router';
 import * as selectors from '../data/selectors';
@@ -8,12 +21,12 @@ import './SessionDetail.scss';
 import { addFavorite, removeFavorite } from '../data/sessions/sessions.actions';
 import { Session } from '../models/Schedule';
 
-interface OwnProps extends RouteComponentProps { };
+interface OwnProps extends RouteComponentProps {}
 
 interface StateProps {
   session?: Session;
-  favoriteSessions: number[],
-};
+  favoriteSessions: number[];
+}
 
 interface DispatchProps {
   addFavorite: typeof addFavorite;
@@ -22,19 +35,23 @@ interface DispatchProps {
 
 type SessionDetailProps = OwnProps & StateProps & DispatchProps;
 
-const SessionDetail: React.FC<SessionDetailProps> = ({ session, addFavorite, removeFavorite, favoriteSessions }) => {
-
+const SessionDetail: React.FC<SessionDetailProps> = ({
+  session,
+  addFavorite,
+  removeFavorite,
+  favoriteSessions,
+}) => {
   if (!session) {
-    return <div>Session not found</div>
+    return <div>Session not found</div>;
   }
 
   const isFavorite = favoriteSessions.indexOf(session.id) > -1;
-  
-  const toggleFavorite = () => { 
+
+  const toggleFavorite = () => {
     isFavorite ? removeFavorite(session.id) : addFavorite(session.id);
   };
-  const shareSession = () => { };
-  const sessionClick = (text: string) => { 
+  const shareSession = () => {};
+  const sessionClick = (text: string) => {
     console.log(`Clicked ${text}`);
   };
 
@@ -47,10 +64,11 @@ const SessionDetail: React.FC<SessionDetailProps> = ({ session, addFavorite, rem
           </IonButtons>
           <IonButtons slot="end">
             <IonButton onClick={() => toggleFavorite()}>
-              {isFavorite ?
-                <IonIcon slot="icon-only" icon={star}></IonIcon> :
+              {isFavorite ? (
+                <IonIcon slot="icon-only" icon={star}></IonIcon>
+              ) : (
                 <IonIcon slot="icon-only" icon={starOutline}></IonIcon>
-              }
+              )}
             </IonButton>
             <IonButton onClick={() => shareSession}>
               <IonIcon slot="icon-only" icon={share}></IonIcon>
@@ -61,8 +79,13 @@ const SessionDetail: React.FC<SessionDetailProps> = ({ session, addFavorite, rem
       <IonContent>
         <div className="ion-padding">
           <h1>{session.name}</h1>
-          {session.tracks.map(track => (
-            <span key={track} className={`session-track-${track.toLowerCase()}`}>{track}</span>
+          {session.tracks.map((track) => (
+            <span
+              key={track}
+              className={`session-track-${track.toLowerCase()}`}
+            >
+              {track}
+            </span>
           ))}
           <p>{session.description}</p>
           <IonText color="medium">
@@ -83,7 +106,12 @@ const SessionDetail: React.FC<SessionDetailProps> = ({ session, addFavorite, rem
           </IonItem>
           <IonItem onClick={() => sessionClick('download video')} button>
             <IonLabel color="primary">Download Video</IonLabel>
-            <IonIcon slot="end" color="primary" size="small" icon={cloudDownload}></IonIcon>
+            <IonIcon
+              slot="end"
+              color="primary"
+              size="small"
+              icon={cloudDownload}
+            ></IonIcon>
           </IonItem>
           <IonItem onClick={() => sessionClick('leave feedback')} button>
             <IonLabel color="primary">Leave Feedback</IonLabel>
@@ -97,11 +125,11 @@ const SessionDetail: React.FC<SessionDetailProps> = ({ session, addFavorite, rem
 export default connect<OwnProps, StateProps, DispatchProps>({
   mapStateToProps: (state, OwnProps) => ({
     session: selectors.getSession(state, OwnProps),
-    favoriteSessions: state.data.favorites
+    favoriteSessions: state.data.favorites,
   }),
   mapDispatchToProps: {
     addFavorite,
-    removeFavorite
+    removeFavorite,
   },
-  component: withRouter(SessionDetail)
-})
+  component: withRouter(SessionDetail),
+});
