@@ -4,28 +4,49 @@ import { RouteComponentProps } from 'react-router';
 import './SpeakerDetail.scss';
 
 import { ActionSheetButton } from '@ionic/core';
-import { IonActionSheet, IonChip, IonIcon, IonHeader, IonLabel, IonToolbar, IonButtons, IonContent, IonButton, IonBackButton, IonPage } from '@ionic/react'
-import { callOutline, callSharp, logoTwitter, logoGithub, logoInstagram, shareOutline, shareSharp } from 'ionicons/icons';
+import {
+  IonActionSheet,
+  IonChip,
+  IonIcon,
+  IonHeader,
+  IonLabel,
+  IonToolbar,
+  IonButtons,
+  IonContent,
+  IonButton,
+  IonBackButton,
+  IonPage,
+} from '@ionic/react';
+import {
+  callOutline,
+  callSharp,
+  logoTwitter,
+  logoGithub,
+  logoInstagram,
+  shareOutline,
+  shareSharp,
+} from 'ionicons/icons';
 
 import { connect } from '../data/connect';
 import * as selectors from '../data/selectors';
 
 import { Speaker } from '../models/Speaker';
 
-
 interface OwnProps extends RouteComponentProps {
   speaker?: Speaker;
-};
+}
 
-interface StateProps {};
+interface StateProps {}
 
-interface DispatchProps {};
+interface DispatchProps {}
 
-interface SpeakerDetailProps extends OwnProps, StateProps, DispatchProps {};
+interface SpeakerDetailProps extends OwnProps, StateProps, DispatchProps {}
 
 const SpeakerDetail: React.FC<SpeakerDetailProps> = ({ speaker }) => {
   const [showActionSheet, setShowActionSheet] = useState(false);
-  const [actionSheetButtons, setActionSheetButtons] = useState<ActionSheetButton[]>([]);
+  const [actionSheetButtons, setActionSheetButtons] = useState<
+    ActionSheetButton[]
+  >([]);
   const [actionSheetHeader, setActionSheetHeader] = useState('');
 
   function openSpeakerShare(speaker: Speaker) {
@@ -34,21 +55,21 @@ const SpeakerDetail: React.FC<SpeakerDetailProps> = ({ speaker }) => {
         text: 'Copy Link',
         handler: () => {
           console.log('Copy Link clicked');
-        }
+        },
       },
       {
         text: 'Share via ...',
         handler: () => {
           console.log('Share via clicked');
-        }
+        },
       },
       {
         text: 'Cancel',
         role: 'cancel',
         handler: () => {
           console.log('Cancel clicked');
-        }
-      }
+        },
+      },
     ]);
     setActionSheetHeader(`Share ${speaker.name}`);
     setShowActionSheet(true);
@@ -60,14 +81,14 @@ const SpeakerDetail: React.FC<SpeakerDetailProps> = ({ speaker }) => {
         text: `Email ( ${speaker.email} )`,
         handler: () => {
           window.open('mailto:' + speaker.email);
-        }
+        },
       },
       {
         text: `Call ( ${speaker.phone} )`,
         handler: () => {
           window.open('tel:' + speaker.phone);
-        }
-      }
+        },
+      },
     ]);
     setActionSheetHeader(`Share ${speaker.name}`);
     setShowActionSheet(true);
@@ -78,7 +99,7 @@ const SpeakerDetail: React.FC<SpeakerDetailProps> = ({ speaker }) => {
   }
 
   if (!speaker) {
-    return <div>Speaker not found</div>
+    return <div>Speaker not found</div>;
   }
 
   return (
@@ -91,36 +112,59 @@ const SpeakerDetail: React.FC<SpeakerDetailProps> = ({ speaker }) => {
             </IonButtons>
             <IonButtons slot="end">
               <IonButton onClick={() => openContact(speaker)}>
-                <IonIcon slot="icon-only" ios={callOutline} md={callSharp}></IonIcon>
+                <IonIcon
+                  slot="icon-only"
+                  ios={callOutline}
+                  md={callSharp}
+                ></IonIcon>
               </IonButton>
               <IonButton onClick={() => openSpeakerShare(speaker)}>
-                <IonIcon slot="icon-only" ios={shareOutline} md={shareSharp}></IonIcon>
+                <IonIcon
+                  slot="icon-only"
+                  ios={shareOutline}
+                  md={shareSharp}
+                ></IonIcon>
               </IonButton>
             </IonButtons>
           </IonToolbar>
         </IonHeader>
 
         <div className="speaker-background">
-          <img src={speaker.profilePic} alt={speaker.name}/>
+          <img src={speaker.profilePic} alt={speaker.name} />
           <h2>{speaker.name}</h2>
         </div>
 
         <div className="ion-padding speaker-detail">
           <p>{speaker.about} Say hello on social media!</p>
 
-          <hr/>
+          <hr />
 
-          <IonChip color="twitter" onClick={() => openExternalUrl(`https://twitter.com/${speaker.twitter}`)}>
+          <IonChip
+            color="twitter"
+            onClick={() =>
+              openExternalUrl(`https://twitter.com/${speaker.twitter}`)
+            }
+          >
             <IonIcon icon={logoTwitter}></IonIcon>
             <IonLabel>Twitter</IonLabel>
           </IonChip>
 
-          <IonChip color="dark" onClick={() => openExternalUrl('https://github.com/ionic-team/ionic')}>
+          <IonChip
+            color="dark"
+            onClick={() =>
+              openExternalUrl('https://github.com/ionic-team/ionic-framework')
+            }
+          >
             <IonIcon icon={logoGithub}></IonIcon>
             <IonLabel>GitHub</IonLabel>
           </IonChip>
 
-          <IonChip color="instagram" onClick={() => openExternalUrl('https://instagram.com/ionicframework')}>
+          <IonChip
+            color="instagram"
+            onClick={() =>
+              openExternalUrl('https://instagram.com/ionicframework')
+            }
+          >
             <IonIcon icon={logoInstagram}></IonIcon>
             <IonLabel>Instagram</IonLabel>
           </IonChip>
@@ -136,10 +180,9 @@ const SpeakerDetail: React.FC<SpeakerDetailProps> = ({ speaker }) => {
   );
 };
 
-
 export default connect({
   mapStateToProps: (state, ownProps) => ({
-    speaker: selectors.getSpeaker(state, ownProps)
+    speaker: selectors.getSpeaker(state, ownProps),
   }),
-  component: SpeakerDetail
+  component: SpeakerDetail,
 });
