@@ -1,5 +1,16 @@
 import React from 'react';
-import { IonHeader, IonToolbar, IonTitle, IonContent, IonPage, IonButtons, IonMenuButton, IonGrid, IonRow, IonCol } from '@ionic/react';
+import {
+  IonHeader,
+  IonToolbar,
+  IonTitle,
+  IonContent,
+  IonPage,
+  IonButtons,
+  IonMenuButton,
+  IonGrid,
+  IonRow,
+  IonCol,
+} from '@ionic/react';
 import SpeakerItem from '../components/SpeakerItem';
 import { Speaker } from '../models/Speaker';
 import { Session } from '../models/Schedule';
@@ -7,19 +18,21 @@ import { connect } from '../data/connect';
 import * as selectors from '../data/selectors';
 import './SpeakerList.scss';
 
-interface OwnProps { };
+interface OwnProps {}
 
 interface StateProps {
   speakers: Speaker[];
   speakerSessions: { [key: string]: Session[] };
-};
+}
 
-interface DispatchProps { };
+interface DispatchProps {}
 
-interface SpeakerListProps extends OwnProps, StateProps, DispatchProps { };
+interface SpeakerListProps extends OwnProps, StateProps, DispatchProps {}
 
-const SpeakerList: React.FC<SpeakerListProps> = ({ speakers, speakerSessions }) => {
-
+const SpeakerList: React.FC<SpeakerListProps> = ({
+  speakers,
+  speakerSessions,
+}) => {
   return (
     <IonPage id="speaker-list">
       <IonHeader translucent={true}>
@@ -38,19 +51,19 @@ const SpeakerList: React.FC<SpeakerListProps> = ({ speakers, speakerSessions }) 
           </IonToolbar>
         </IonHeader>
 
-          <IonGrid fixed>
-            <IonRow>
-              {speakers.map(speaker => (
-                <IonCol size="12" size-md="6" key={speaker.id}>
-                  <SpeakerItem
-                    key={speaker.id}
-                    speaker={speaker}
-                    sessions={speakerSessions[speaker.name]}
-                  />
-                </IonCol>
-              ))}
-            </IonRow>
-          </IonGrid>
+        <IonGrid fixed>
+          <IonRow>
+            {speakers.map((speaker) => (
+              <IonCol size="12" size-md="6" key={speaker.id}>
+                <SpeakerItem
+                  key={speaker.id}
+                  speaker={speaker}
+                  sessions={speakerSessions[speaker.name]}
+                />
+              </IonCol>
+            ))}
+          </IonRow>
+        </IonGrid>
       </IonContent>
     </IonPage>
   );
@@ -59,7 +72,7 @@ const SpeakerList: React.FC<SpeakerListProps> = ({ speakers, speakerSessions }) 
 export default connect<OwnProps, StateProps, DispatchProps>({
   mapStateToProps: (state) => ({
     speakers: selectors.getSpeakers(state),
-    speakerSessions: selectors.getSpeakerSessions(state)
+    speakerSessions: selectors.getSpeakerSessions(state),
   }),
-  component: React.memo(SpeakerList)
+  component: React.memo(SpeakerList),
 });
